requirements-dev.txt
pytest==2.9.2
pytest-mock==1.5.0
pytest-cov==2.4.0
requests-mock==1.3.0
pytest-catchlog==1.2.2
pylint==1.5.4
webtest==2.0.27

